Allow custom summaries
This is an "escape hatch" that can be used when neither automatic nor manual summary splitting is sufficient to create a good summary.  This is specifically useful for articles that begin with a disclaimer or anything else that shouldn't be included in the summary.  It's also useful for footnote links and other mid-text items that should be excluded from the summary.